+++ pygbif/occurrences/get.py
@@ -50,8 +50,8 @@ def get_fragment(key, **kwargs):
 
         from pygbif import occurrences
         occurrences.get_fragment(key = 1052909293)
-        occurrences.get_fragment(key = 1052909293)
-        occurrences.get_fragment(key = 1052909293)
+        occurrences.get_fragment(key = 1227768771)
+        occurrences.get_fragment(key = 1227769518)
     '''
     url = gbif_baseurl + 'occurrence/' + str(key) + '/fragment'
     out = gbif_GET(url, {}, **kwargs)