+++ pygbif/registry/__init__.py
@@ -4,3 +4,4 @@ from .nodes import nodes
 from .networks import networks
 from .installations import installations
 from .datasets import datasets, dataset_metrics
+from .organizations import organizations