+++ Changelog.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.4.0 (2019-11-20)
+------------------
+- changed base url to https for all requests; was already https for maps and downloads in previous versions
+- occurrences, species, and registry modules gain docstrings with brief summary of each method
+- pygbif gains ability to cache http requests. caching is off by default. See `?pygbif.caching` for all the details :issue:`52` :issue:`56` via @nleguillarme
+- made note in docs that if you are trying to get the same behavior as the GBIF website for name searching, `species.name_backbone` is likely what you want :issue:`55` thanks @qgroom
+
 0.3.0 (2019-01-25)
 ------------------
 - pygbif is Python 3 only now :issue:`19`